Cloud Services/GCP_Analysis1-machine_learning/machine-learning-cloud-function/requirements.txt
google-api-core==1.22.0
google-auth==1.19.2
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
sklearn==0.0
scikit-learn==0.22.2.post1
